versione 2 luglio

+++ demo/src/main/java/com/example/demo/projection/ExclusionRulesProjection.java
@@ -6,6 +6,6 @@ public interface ExclusionRulesProjection {
 	BigInteger getId();         
     String getCode();
     String getName();
-    Integer getCompanyCount();
+    Long getCompanyCount();
     String getStatus();    
 }
